labela_backend_assignment/autocompany/serializers.py
from dataclasses import field, fields
from pyexpat import model
from statistics import mode
from rest_framework import serializers
from api.models import Cart, Order, Part
from users.models import MyUser as User

class PartSerializer(serializers.ModelSerializer):
    class Meta:
        model = Part
        fields = ['id', 'name', 'description', 'price']


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'name', 'email', 'password', 'username']

class CartSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cart
        fields = ['id', 'user_id', 'parts']

class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = ['id', 'user_id', 'parts', 'delivery_time']
