labela_backend_assignment/requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
djangorestframework==3.13.1
django-admin==2.0.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0